Enhance HTTP server task and utilities with improved functionality and documentation

+++ src/http-server/http_server_task.c
@@ -3,14 +3,16 @@
  * @brief HTTP Server FreeRTOS Task Implementation
  */
 
+#include "http_server_task.h"
+
 #include <stdint.h>
 #include <stdio.h>
-#include <limits.h>
 
 #include "FreeRTOS.h"
 #include "task.h"
 
 #include "pico/stdlib.h"
+
 #include "system_config.h"
 #include "network_task.h"
 #include "logger.h"
@@ -46,7 +48,7 @@ static bool wait_for_network_ready(uint32_t timeout_ms)
         uint32_t notification_value = 0;
         
         // Wait for notification with timeout
-        if (xTaskNotifyWait(0, ULONG_MAX, &notification_value, pdMS_TO_TICKS(100)) == pdTRUE)
+        if (xTaskNotifyWait(0, UINT32_MAX, &notification_value, pdMS_TO_TICKS(100)) == pdTRUE)
         {
             if (notification_value & NETWORK_NOTIFY_READY)
             {
@@ -92,7 +94,7 @@ static void vHttpServerTask(void *pvParameters)
         uint32_t notification_value = 0;
         
         // Check for network status changes (non-blocking)
-        if (xTaskNotifyWait(0, ULONG_MAX, &notification_value, 0) == pdTRUE)
+        if (xTaskNotifyWait(0, UINT32_MAX, &notification_value, 0) == pdTRUE)
         {
             if (notification_value & NETWORK_NOTIFY_NOT_READY)
             {
@@ -131,9 +133,6 @@ static void vHttpServerTask(void *pvParameters)
     }
 }
 
-/**
- * @brief Initialize HTTP server
- */
 /**
  * @brief Initialize and create the HTTP server task
  */
